src/main/java/webservices/httpclient/ShoppingBagClient.java
package webservices.httpclient;

import com.mashape.unirest.http.HttpResponse;
import com.mashape.unirest.http.JsonNode;
import com.mashape.unirest.http.Unirest;
import com.mashape.unirest.http.exceptions.UnirestException;

public class ShoppingBagClient extends RestClient {

    public ShoppingBagClient() {
        super();
    }


    @Override
    protected void setRessourceName() {
        this.ressource = "shopping-bags";
    }

    public HttpResponse<JsonNode> getItemById(Integer itemId) throws UnirestException {
        return Unirest
                .get(buildItemURI(itemId))
                .asJson();
    }
}

